A bit of package restructuring
	added localstack.http.resources.swagger
	all the swagger stuff in there
