+++ localstack-core/localstack/http/resources/swagger/static/favicon-16x16.png
