+++ localstack-core/localstack/http/resources/swagger/static/favicon-32x32.png
