+++ localstack-core/localstack/http/resources/swagger/static/index.css
