+++ localstack-core/localstack/http/resources/swagger/static/oauth2-redirect.html
