+++ localstack-core/localstack/http/resources/swagger/static/swagger-initializer.js
