+++ localstack-core/localstack/http/resources/swagger/static/swagger-ui-standalone-preset.js
