+++ localstack-core/localstack/http/resources/swagger/static/swagger-ui.js
