+++ localstack-core/localstack/http/resources/swagger/templates/index.html
