+++ localstack-core/localstack/plugins.py
@@ -1,20 +1,10 @@
-import importlib
 import logging
-
-import werkzeug
-import yaml
-from rolo.routing import RuleAdapter
 
 from localstack import config
 from localstack.aws.handlers.validation import OASPlugin
-from localstack.http import Response
 from localstack.runtime import hooks
-from localstack.services.edge import ROUTER
-from localstack.services.internal import get_internal_apis
 from localstack.utils.files import rm_rf
-from localstack.utils.openapi import get_localstack_openapi_spec
 from localstack.utils.ssl import get_cert_pem_file_path
-from localstack.utils.swagger import SwaggerUIApi
 
 LOG = logging.getLogger(__name__)
 
@@ -32,25 +22,6 @@ def delete_cached_certificate():
     LOG.debug("Removing the cached local SSL certificate")
     target_file = get_cert_pem_file_path()
     rm_rf(target_file)
-
-
-@hooks.on_infra_start()
-def register_swagger_endpoints():
-    get_internal_apis().add(SwaggerUIApi())
-
-    def _serve_static_file(_request, path: str):
-        module = importlib.import_module("localstack.static")
-        return Response.for_resource(module, path)
-
-    def _serve_openapi_spec(_request):
-        spec = get_localstack_openapi_spec()
-        response_body = yaml.dump(spec)
-        return werkzeug.Response(
-            response_body, content_type="application/yaml", direct_passthrough=True
-        )
-
-    ROUTER.add(RuleAdapter("/openapi.yaml", _serve_openapi_spec))
-    ROUTER.add(RuleAdapter("/static/<path:path>", _serve_static_file))
 
 
 class CoreOASPlugin(OASPlugin):
